@ajaybhargavb 2016-07-14
Perf Data:

Before:

                            Method |        Median |     StdDev |
---------------------------------- |-------------- |----------- |
                         CreateRvd |    65.8707 ns |  1.6964 ns |
           CreateRvd_SetProp_Clear |   125.9218 ns |  4.1885 ns |
 CreateRvd_SetMultipleProps_Remove | 1,212.2968 ns | 28.6963 ns |



After:

                            Method |        Median |     StdDev |
---------------------------------- |-------------- |----------- |
                         CreateRvd |    60.2613 ns |  1.1405 ns |
           CreateRvd_SetProp_Clear |    92.1897 ns | 22.6009 ns |
 CreateRvd_SetMultipleProps_Remove | 1,317.1857 ns | 33.5729 ns |