@rynowak 2016-07-14
Let's review the perf numbers when you get in, I'd like to understand what's in these tests.