@ajaybhargavb 2016-07-15
:up: :date: 

Added a test. Didn't add other tests for `RemoveAt` and `Clear` because there are tests already covering that code path.