@ajaybhargavb 2016-07-18
:up: :date: 
Added a test for `Remove`.